Start adding tests for the annotation layer.
Split the annotations into their own source file.  This could later become a directory of the same name (annotation) with each annotation in a separate file, if needed.
